# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Masal — isolated portals and implemented v1 system API
  version: 0.7.0
  description: |
    Source snapshot 07 includes the implemented Laravel v1 modules listed below.
    This document retains the 35 paths / 45 detailed operations originally described
    in the 0.3 accounts/catalog contract. Those detailed schemas are a limited baseline,
    not complete request/response coverage for snapshot 07. Subsequent implementation
    behavior is authoritative in each current FormRequest, controller, resource and service.
    API_ROUTE_CATALOG.json is generated from the current Laravel route:list --json output;
    it lists all 214 registered api/v1 and Sanctum CSRF method/path records (180 unique
    URI paths; GET records also include HEAD). The catalog proves registered path coverage,
    not permissions, business correctness, external-provider readiness or production deployment.
    Validation status and remaining native/deployment work are recorded in
    ../docs/FULL_SYSTEM_07_VERIFICATION.md. Production is still release 06b at this draft.
servers:
- url: https://admin.dananir-iq.com
  description: Final System administration portal; same-origin cookie session and /api/v1 paths
- url: https://agents.dananir-iq.com
  description: Final main-agent, sub-agent and sub-branch portal; same-origin /api/v1 paths
- url: https://pos.dananir-iq.com
  description: Final point-of-sale portal; same-origin /api/v1 paths
- url: /
...
paths:
...
  /api/v1/catalog/products:
    get:
      tags:
      - Catalog
      summary: عرض الفئات
      description: صلاحية products.view مطلوبة. عرض الفئات ضمن الشبكة المسموحة؛ principal يبدأ بقائمة
        فارغة حتى التعيين، وقواعد الأسلاف تتقاطع. كل mutation للكتالوج العالمي مقصور على حساب النظام أو
        موظف يشمل نطاقه النظام.
      security: &id001
      - AdminSession: []
      - AgentsSession: []
      - PosSession: []
      parameters:
      - $ref: '#/components/parameters/LocalPortal'
      - &id003
        in: query
        name: page
        schema:
          type: integer
          minimum: 1
      - &id004
        in: query
        name: per_page
        schema:
          type: integer
          minimum: 1
          maximum: 100
      - &id005
        in: query
        name: query
        schema:
          type: string
          maxLength: 160
      - &id006
        in: query
        name: provider
        schema:
          type: integer
          minimum: 1
      - &id007
        in: query
        name: kind
        schema:
          type: string
          enum:
          - محلية
          - عالمية
      - &id008
        in: query
        name: state
        schema:
          type: string
          enum:
          - active
          - inactive
      - &id009
        in: query
        name: supplier
        schema:
          type: string
          maxLength: 160
      - &id010
        in: query
        name: connection
        schema:
          type: string
          enum:
          - ملفات
          - API
      - &id011
        in: query
        name: from
        schema:
          type: string
          format: date
      - &id012
        in: query
        name: to
        schema:
          type: string
          format: date
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogProductCollection'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailure'
    post:
      tags:
      - Catalog
      summary: إضافة الفئات
      description: 'صلاحية products.view مطلوبة. عرض الفئات ضمن الشبكة المسموحة؛ principal يبدأ بقائمة
        فارغة حتى التعيين، وقواعد الأسلاف تتقاطع. كل mutation للكتالوج العالمي مقصور على حساب النظام أو
        موظف يشمل نطاقه النظام. صور وإزالة: صلاحية products.images. إرسال image وremove_image معًا مرفوض.'
      security: *id001
      parameters:
      - $ref: '#/components/parameters/LocalPortal'
      - $ref: '#/components/parameters/XsrfToken'
      responses:
        '201':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogProductResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailure'
        '409':
          $ref: '#/components/responses/Conflict'
        '419':
          $ref: '#/components/responses/CsrfFailure'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogProductCreate'
          multipart/form-data:
            schema:
              type: object
              properties:
                payload:
                  type: string
                  description: JSON مطابق لـCreate أو Patch بحسب العملية، بلا id أو status.
                image:
                  type: string
                  format: binary
                  description: PNG/JPEG حتى 700000 بايت وأبعاد 4096×4096؛ حفظ خاص ضمن نفس المعاملة.
                _method:
                  type: string
                  enum:
                  - PATCH
                  description: يرسل عند تعديل multipart عبر POST إلى مسار السجل.
              required:
              - payload
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: masal_admin_session
      description: Browser-managed HttpOnly cookie issued only by the admin portal
    AgentsSession:
      type: apiKey
      in: cookie
      name: masal_agents_session
      description: Browser-managed HttpOnly cookie issued only by the agents portal
    PosSession:
      type: apiKey
      in: cookie
      name: masal_pos_session
      description: Browser-managed HttpOnly cookie issued only by the POS portal
  parameters:
    LocalPortal:
      name: X-Masal-Portal
      in: header
      required: false
      description: 'فقط عند التطوير/الاختبار على مضيف محلي مشترك. في الإنتاج لا يحدد

        هذا الحقل البوابة ولا يمنح وصولًا؛ يعتمد السيرفر المضيف المحدد في إعداداته.

        '
      schema:
        type: string
        enum:
        - admin
        - agents
        - pos
    XsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: false
      description: Client always sends the URL-decoded current host XSRF-TOKEN cookie; Laravel 13 uses
        token validation as fallback when same-origin verification does not pass
      schema:
        type: string
...
  schemas:
...
    Error:
      type: object
      required:
      - message
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required:
      - message
      - errors
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    CatalogProduct:
      type: object
      properties:
        id: &id015
          type: integer
        name: &id016
          type: string
          maxLength: 160
        provider_id: &id017
          type: integer
          minimum: 1
        provider_name: &id018
          type: string
        provider_status:
          type: string
          enum:
          - active
          - disabled
        provider_logo_url:
          type: string
          nullable: true
        kind:
          type: string
          enum:
          - محلية
          - عالمية
        currency: &id019
          type: string
          enum:
          - IQD
          - USD
        face_value: &id014
          type: string
          pattern: ^\d+(\.\d{1,2})?$
          nullable: true
          description: قيمة decimal كنص؛ null يعني غير محددة في قاموس الأصل، وليس صفرًا أو تكلفة مفترضة.
        minimum_price: *id014
        daily_limit_type:
          type: string
          enum:
          - ''
          - quantity
          - amount
          description: الفئات المرجعية المستوردة تبدأ بلا حد؛ الحفظ الجديد يلزم quantity أو amount.
        daily_quantity:
          type: integer
          nullable: true
          minimum: 1
        daily_amount: *id014
        field_policy:
          type: object
          properties:
            pin:
              type: string
              enum:
              - required
            expiry:
              type: string
              enum:
              - required
            serial:
              type: string
              enum:
              - required
              - unused
            cvc:
              type: string
              enum:
              - required
              - unused
            reference:
              type: string
              enum:
              - required
              - unused
          required:
          - pin
          - expiry
          - serial
          - cvc
          - reference
          additionalProperties: false
        extra_fields:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
                pattern: ^[a-z][a-z0-9_]{1,30}$
              label:
                type: string
                maxLength: 160
              required:
                type: boolean
            required:
            - key
            - label
            - required
          maxItems: 30
        display_order:
          type: integer
          minimum: 0
          maximum: 1000000
        import_codes:
          type: array
          items:
            type: string
            maxLength: 80
          maxItems: 50
          uniqueItems: true
        receipt_language:
          type: string
          enum:
          - ''
          - ar
          - en
          - ku
        receipt_width:
          type: integer
          enum:
          - 58
          - 80
        receipt_header:
          type: string
          maxLength: 1000
        receipt_footer:
          type: string
          maxLength: 1000
        allowed_cities:
          type: array
          items:
            type: string
          maxItems: 19
          uniqueItems: true
        status: &id020
          type: string
          enum:
          - active
          - disabled
        image_url:
          type: string
          nullable: true
        version: *id013
        created_at:
          type: string
          format: date-time
          nullable: true
      required:
      - id
      - name
      - provider_id
      - provider_name
      - provider_status
      - provider_logo_url
      - kind
      - currency
      - face_value
      - minimum_price
      - daily_limit_type
      - daily_quantity
      - daily_amount
      - field_policy
      - extra_fields
      - display_order
      - import_codes
      - receipt_language
      - receipt_width
      - receipt_header
      - receipt_footer
      - allowed_cities
      - status
      - image_url
      - version
      - created_at
...
    CatalogPageMeta:
      type: object
      properties:
        current_page:
          type: integer
          minimum: 1
        last_page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
      required:
      - current_page
      - last_page
      - per_page
      - total
    CatalogProductCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 160
        provider_id:
          type: integer
          minimum: 1
        kind:
          type: string
          enum:
          - محلية
          - عالمية
        currency:
          type: string
          enum:
          - IQD
          - USD
        face_value:
          oneOf:
          - type: number
            minimum: 1
            maximum: 999999999999.99
          - type: string
            pattern: ^\d+(\.\d{1,2})?$
          description: رقم أو نص عشري موجب، منزلتان كحد أقصى.
        minimum_price:
          oneOf:
          - type: number
            minimum: 1
            maximum: 999999999999.99
          - type: string
            pattern: ^\d+(\.\d{1,2})?$
          description: رقم أو نص عشري موجب، منزلتان كحد أقصى.
        daily_limit_type:
          type: string
          enum:
          - quantity
          - amount
        daily_quantity:
          type: integer
          nullable: true
          minimum: 1
        daily_amount:
          oneOf:
          - type: number
            minimum: 0.01
            maximum: 999999999999.99
          - type: string
            pattern: ^\d+(\.\d{1,2})?$
          nullable: true
        field_policy:
          type: object
          properties:
            pin:
              type: string
              enum:
              - required
            expiry:
              type: string
              enum:
              - required
            serial:
              type: string
              enum:
              - required
              - unused
            cvc:
              type: string
              enum:
              - required
              - unused
            reference:
              type: string
              enum:
              - required
              - unused
          required:
          - pin
          - expiry
          - serial
          - cvc
          - reference
          additionalProperties: false
        extra_fields:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
                pattern: ^[a-z][a-z0-9_]{1,30}$
              label:
                type: string
                maxLength: 160
              required:
                type: boolean
            required:
            - key
            - label
            - required
          maxItems: 30
        display_order:
          type: integer
          minimum: 0
          maximum: 1000000
        import_codes:
          type: array
          items:
            type: string
            maxLength: 80
          maxItems: 50
          uniqueItems: true
        receipt_language:
          type: string
          enum:
          - ''
          - ar
          - en
          - ku
          nullable: true
        receipt_width:
          type: integer
          enum:
          - 58
          - 80
        receipt_header:
          type: string
          maxLength: 1000
        receipt_footer:
          type: string
          maxLength: 1000
        allowed_cities:
          type: array
          items:
            type: string
          maxItems: 19
          uniqueItems: true
        remove_image:
          type: boolean
        reason:
          type: string
          nullable: true
          maxLength: 2000
      required:
      - name
      - provider_id
      - kind
      - face_value
      - currency
      - minimum_price
      - daily_limit_type
      - field_policy
      - extra_fields
      - display_order
      - import_codes
      - receipt_width
      - allowed_cities
      additionalProperties: false
...
    CatalogProductResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/CatalogProduct'
      required:
      - data
    CatalogProductCollection:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CatalogProduct'
        meta:
          $ref: '#/components/schemas/CatalogPageMeta'
        links:
          type: object
          properties: {}
      required:
      - data
      - meta
...
  responses:
    Unauthenticated:
      description: Missing, expired, or invalid session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Unknown host, portal mismatch, disabled session account chain, or insufficient permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    CsrfFailure:
      description: Origin verification did not pass and CSRF token is missing or invalid; initialize a
        fresh CSRF session before retrying login
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationFailure:
      description: Invalid input or generic rejected login; no account existence or portal reason is revealed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NotFound:
      description: Account does not exist or is outside the authenticated account scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Stale optimistic version, or a linked permission profile that cannot be deleted. No
        requested mutation is committed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'